Get rid of in-block allocation for variable size memory allocation blocks.
The issue is that the memory freeing algorithm had to check the low byte of the info block *before* current block, and in case the allocation block sits at the beginning of a memory page that could cause a page fault.

By always allocating after the current block we avoid this issue.